--- client/src/components/pages/Profile/notes.html
@@ -1,13 +1,0 @@
-old user 
-old planets 
-their zone 
-search bar (planets)
-filter (color, type, sunlight, water, eatable)
-recommend plants from location 
-
-
-new user
-search bar (planets)
-filter (color, type, sunlight, water, eatable)
-search bar (location)
-recommend plants from location
