+++ client/src/components/pages/Search/SearchModal.css
@@ -14,8 +14,9 @@
     background: white;
     padding: 20px;
     border-radius: 5px;
-    width: 80%;
-    max-width: 600px;
+    width: 80vw;
+    height: 80vh;
+    /* max-width: 600px; */
     position: relative;
 }
 
@@ -23,6 +24,16 @@
     position: absolute;
     top: 10px;
     right: 10px;
+    background: transparent;
+    border: none;
+    font-size: 18px;
+    cursor: pointer;
+}
+
+.add-button {
+    position: absolute;
+    top: 10px;
+    right: 100px;
     background: transparent;
     border: none;
     font-size: 18px;
